CromSerial: sleep before/after setting each valve position
The previous code worked fine before, but it no longer works, I don't
know why. Now, with the code before this commit, when switching from
00000 to 11111, some valves are not switched, the same when closing the
valves.

The obvious thing to try to fix the problem is to add a small delay
between each writing to the serial port; and it works! Problem solved!
(I hope… With electronics we are never sure.)